+++ core/project.go
@@ -16,6 +16,7 @@ const (
 var (
 	ErrProjectNotFound      = errors.New("project not found")
 	ErrProjectAlreadyExists = errors.New("project already exists")
+	ErrParentlessModule     = errors.New("no parent project for module exists, please create parent first")
 )
 
 type Project struct {
@@ -45,6 +46,30 @@ func (t *Timetrace) LoadProject(key string) (*Project, error) {
 	return t.loadProject(path)
 }
 
+// ListProjectModules loads all modules for a project and returns their keys as a concatenated string
+func (t *Timetrace) ListProjectModules(project *Project) (string, error) {
+	allModules, err := t.loadProjectModules(project)
+	if err != nil {
+		return "", err
+	}
+
+	if len(allModules) == 0 {
+		return "-", nil
+	}
+
+	var mList string
+	for i, p := range allModules {
+		// get the name of the module without the prefix
+		mList += strings.Split(p.Key, "@")[0]
+		// append comma if it is not the last element
+		if i+1 != len(allModules) {
+			mList += ","
+		}
+	}
+
+	return mList, nil
+}
+
 // ListProjects loads and returns all stored projects sorted by their filenames.
 // If no projects are found, an empty slice and no error will be returned.
 func (t *Timetrace) ListProjects() ([]*Project, error) {
@@ -69,9 +94,15 @@ func (t *Timetrace) ListProjects() ([]*Project, error) {
 // SaveProject persists the given project. Returns ErrProjectAlreadyExists if
 // the project already exists and saving isn't forced.
 func (t *Timetrace) SaveProject(project Project, force bool) error {
-	path := t.fs.ProjectFilepath(project.Key)
+	if project.IsModule() {
+		err := t.assertParent(project)
+		if err != nil {
+			return err
+		}
+	}
 
-	if _, err := os.Stat(path); os.IsExist(err) && !force {
+	path := t.fs.ProjectFilepath(project.Key)
+	if _, err := os.Stat(path); err == nil && !force {
 		return ErrProjectAlreadyExists
 	}
 
@@ -151,7 +182,7 @@ func (t *Timetrace) loadProjectModules(project *Project) ([]*Project, error) {
 
 	for _, p := range projects {
 		if p.Parent() == project.Key {
-			modules = append(modules, project)
+			modules = append(modules, p)
 		}
 	}
 
@@ -168,4 +199,19 @@ func (t *Timetrace) editorFromEnvironment() string {
 	}
 
 	return defaultEditor
+}
+
+func (t *Timetrace) assertParent(project Project) error {
+	allP, err := t.ListProjects()
+	if err != nil {
+		return err
+	}
+
+	for _, p := range allP {
+		if p.Key == project.Parent() {
+			return nil
+		}
+	}
+
+	return ErrParentlessModule
 }